Hide `Add Task` button if current relation is empty to fix #128
I have used `@ViewBuilder` attribute to pass the appropriate `View` to the `navigationBarItems(trailing:)` method. Before collecting data from the server the `id` of `currentRelation` is 0 and if the user doesn't have any Relation then the server response for `id` would be nil. This is why I have checked both of the conditions.